+++ lib/actions/transformSchemaToWidget.js
@@ -133,9 +133,12 @@ const processSchema = (fileSchema) => {
   if (isModule) {
     fileSchema.moduleProps = {
       name: fileSchema.filePath,
-      // INFO: isso será feito no final para usar o final bundle
+      // INFO: isso será trocado no final para usar o final bundle
       module: wrapCodeInGetFunction(fileSchema.content, fileSchema.filePath),
-      values: getFunctionExportDeclarationKeys(fileSchema.content),
+      values: getFunctionExportDeclarationKeys(
+        fileSchema.content,
+        fileSchema.filePath,
+      ),
       // values: extractTopLevelDeclarations(
       //   fileSchema.pureJsContent,
       //   fileSchema.filePath,
@@ -674,7 +677,14 @@ const prepareListOfInjections = (fileSchemas, fileSchema) => {
   return fileSchema;
 };
 
-const foo = (fileSchemas, fileSchema) => {
+/**
+ * Injeta a referencia dos módulos no topo de cada arquivo final que depende dos módulos
+ * ou que tem dependentes que dependem dos módulos
+ * @param {*} fileSchemas
+ * @param {*} fileSchema
+ * @returns
+ */
+const injectModules = (fileSchemas, fileSchema) => {
   let fileBundle = fileSchema.finalFileBundle;
 
   // Se nao tiver nada a ser injetado, somente retorna o file Schema sem alterações
@@ -696,9 +706,10 @@ const foo = (fileSchemas, fileSchema) => {
 
     let injections = "";
 
+    // m = modulesCode
     moduleValuesKeys.forEach((propKey) => {
       injections = `
-           const ${propKey} = props.alem.modulesCode["${fileItemSchema.moduleProps.name}"]().${propKey};
+           const ${propKey} = props.alem.m["${fileItemSchema.moduleProps.name}"]().${propKey};
            ${injections}
            `;
     });
@@ -714,7 +725,7 @@ const foo = (fileSchemas, fileSchema) => {
 
   // Se o arquivo for um módulo...
   if (fileSchema.isModule) {
-    console.log("FOOO", fileSchema.filePath);
+    // console.log("FOOO", fileSchema.filePath);
     // Atualiza o modulo deste arquivo para usar o finalBundle com as importacoes injetadas
     fileSchema.moduleProps = {
       ...fileSchema.moduleProps,
@@ -729,7 +740,15 @@ const foo = (fileSchemas, fileSchema) => {
   return fileSchema;
 };
 
-const injectDependencies = (fileSchemas, fileSchema, skipModules = true) => {
+/**
+ * Coloca o conteúdo dos arquivos nao .ts e .jsx de dependencia dentro do bundle de cada arquivo do schema global
+ * Esse é um processo que ocorre para todos os arquivos, mas somente copia e cola o conteudo para arquivos nao JSX.
+ *
+ * Arquivos reconhecidos como JSX (Widgets) serão tratados de outra forma. Ver "swapComponentsForWidgets"
+ * @param {{filePath: string, toImport: string[], content: string, finalFileBundle: string, componentImportItems:[], toBeInjectedFiles:[], componentParamsItems:[], componentComponentItems: [], widgetName?: string, htmlElementsProps: {}}[]} fileSchemas
+ * @param {{filePath: string, toImport: string[], content: string, finalFileBundle: string, componentImportItems:[], toBeInjectedFiles:[], componentParamsItems:[], componentComponentItems: [], widgetName?: string, htmlElementsProps: {}}} fileSchema
+ */
+const injectFilesDependencies = (fileSchemas, fileSchema) => {
   let fileBundle = fileSchema.finalFileBundle;
 
   // Se nao tiver nada a ser injetado, somente retorna o file Schema sem alterações
@@ -764,28 +783,6 @@ const injectDependencies = (fileSchemas, fileSchema, skipModules = true) => {
 
   fileSchema.finalFileBundle = fileBundle;
   fileSchema.injectedFiles = injectedFiles;
-
-  return fileSchema;
-};
-
-/**
- * Coloca o conteúdo dos arquivos nao .ts e .jsx de dependencia dentro do bundle de cada arquivo do schema global
- * Esse é um processo que ocorre para todos os arquivos, mas somente copia e cola o conteudo para arquivos nao JSX.
- *
- * Arquivos reconhecidos como JSX (Widgets) serão tratados de outra forma. Ver "swapComponentsForWidgets"
- * @param {{filePath: string, toImport: string[], content: string, finalFileBundle: string, componentImportItems:[], toBeInjectedFiles:[], componentParamsItems:[], componentComponentItems: [], widgetName?: string, htmlElementsProps: {}}[]} fileSchemas
- * @param {{filePath: string, toImport: string[], content: string, finalFileBundle: string, componentImportItems:[], toBeInjectedFiles:[], componentParamsItems:[], componentComponentItems: [], widgetName?: string, htmlElementsProps: {}}} fileSchema
- */
-const injectFilesDependencies = (fileSchemas, fileSchema) => {
-  fileSchema = injectDependencies(fileSchemas, fileSchema, true);
-
-  return fileSchema;
-};
-
-const injectModulesDependencies = (fileSchemas, fileSchema) => {
-  // if (fileSchema.isModule) {
-  //   fileSchema = injectDependencies(fileSchemas, fileSchema, false);
-  // }
 
   return fileSchema;
 };
@@ -936,72 +933,13 @@ const transformSchemaToWidget = (fileSchemas, additionalFileSchemas) => {
     fileSchemas[fileSchemaIndex] = fileSchema;
   });
 
-  // Injeta as dependencias dos módulos
-  fileSchemas.forEach((fileSchema, fileSchemaIndex) => {
-    fileSchemas[fileSchemaIndex] = injectModulesDependencies(
-      fileSchemas,
-      fileSchema,
-    );
-  });
-
-  // FOO
   // ATENÇÃO: muda tanto o "finalFileBundle" quanto o "jsContent"
+  // Injeta os módulos caso sejam requeridos
   fileSchemas.forEach((fileSchema, fileSchemaIndex) => {
-    fileSchemas[fileSchemaIndex] = foo(fileSchemas, fileSchema);
-    // if (fileSchema.filePath.includes("src/Main.tsx")) {
-    //   console.log("========== AAAA");
-    //   console.log("fileSchemas[fileSchemaIndex]", fileSchemas[fileSchemaIndex]);
-    // }
+    fileSchemas[fileSchemaIndex] = injectModules(fileSchemas, fileSchema);
   });
-
-  // Ajusta módulos: Faz o parse de caminho das dependencias dos módulos que importam outros módulos
-  // fileSchemas.forEach((fileSchema, fileSchemaIndex) => {
-  //   fileSchemas[fileSchemaIndex] = parseModules(fileSchemas, fileSchema);
-  // });
 
   return { fileSchemas, processError };
 };
-
-/**
- * Faz a troca de referencia das declarações para apontarem para o props.alem.modulesCode dos outros módulos
- * @param {*} fileSchemas
- * @param {*} fileSchema
- */
-// const parseModules = (fileSchemas, fileSchema) => {
-//   // Se for um módulo && esse módulo tiver importando outros módulos...
-//   if (fileSchema.isModule && fileSchema.toBeInjectedModules) {
-//     fileSchema.toBeInjectedModules.forEach((modulePath) => {
-//       const moduleSchema = fileSchemas.find((m) => m.filePath === modulePath);
-
-//       const moduleSchemaItems = Object.keys(moduleSchema.moduleProps.values);
-//       const fileSchemaModuleItems = Object.keys(fileSchema.moduleProps.values);
-//       moduleSchemaItems.forEach((moduleItemKey) => {
-//         // 1 - Neste ponto temos uma referencia/chave do módulo importado e do modulo atual
-//         // 2 - Agora faz um loop para checar e tratar referencia em cada item do modulo atual (fileSchema)
-//         fileSchemaModuleItems.forEach((currentModuleKey) => {
-//           const currentModuleItemValue =
-//             fileSchema.moduleProps.values[currentModuleKey];
-//           // Se o item atual do modulo atual possuir qualquer referencia de um item do modulo sendo importando
-//           // aplica a mudanca de referencia
-//           if (currentModuleItemValue.includes(moduleItemKey)) {
-//             const replaced = transformVariableInCode(
-//               currentModuleItemValue,
-//               moduleItemKey,
-//               ":::VAR_REF:::",
-//             );
-//             // Atualiza as referencias no fileSchema (modulo) atual
-//             fileSchema.moduleProps.values[currentModuleKey] =
-//               replaced.replaceAll(
-//                 ":::VAR_REF:::",
-//                 `props.alem.modulesCode['${moduleSchema.moduleProps.name}'].${moduleItemKey}`,
-//               );
-//           }
-//         });
-//       });
-//     });
-//   }
-
-//   return fileSchema;
-// };
 
 module.exports = transformSchemaToWidget;
