+++ lib/parsers/extractTopLevelDeclarations.js
@@ -133,9 +133,10 @@ const processDeclarations = (declarations, modulePath) => {
           ":::VAR_REF:::",
         );
 
+        // m = modulesCode
         declarations[refKey] = replaced.replaceAll(
           ":::VAR_REF:::",
-          `props.alem.modulesCode['${modulePath}'].${declarationReference}`,
+          `props.alem.m['${modulePath}'].${declarationReference}`,
         );
       }
     });
