+++ lib/parsers/getFunctionExportDeclarationKeys.js
@@ -5,8 +5,11 @@ const transformAndGetFunctionExportDeclarations = require("./transformAndGetFunc
  * @param {*} code
  * @returns
  */
-const getFunctionExportDeclarationKeys = (code) => {
-  const declarations = transformAndGetFunctionExportDeclarations(code);
+const getFunctionExportDeclarationKeys = (code, filePath) => {
+  const declarations = transformAndGetFunctionExportDeclarations(
+    code,
+    filePath,
+  );
   // console.log("RESULTADO:");
   // console.log(declarations);
   // console.log("\n");
