+++ lib/parsers/transformAndGetFunctionExportDeclarations.js
@@ -1,6 +1,7 @@
 const parser = require("@babel/parser");
 const traverse = require("@babel/traverse").default;
 const t = require("@babel/types");
+const convertFunctionsToArrow = require("./convertFunctionsToArrow");
 const generate = require("@babel/generator").default;
 
 function convertExportDefault(code) {
@@ -33,60 +34,74 @@ function convertExportDefault(code) {
  * @param {*} code
  * @returns
  */
-function transformAndGetFunctionExportDeclarations(code) {
-  // code = code.replaceAll(/\bexport default { \b/g, "export");
+function transformAndGetFunctionExportDeclarations(code, filePath) {
+  // Transforma "export default function" para "function"
+  code = code.replaceAll(/\bexport default function\b/g, "function");
+
+  // Converte functions para arrow functions
+  try {
+    code = convertFunctionsToArrow(code);
+  } catch (error) {
+    console.log(code);
+    throw new Error(`File: ${filePath}: ${error}`);
+  }
+
   code = convertExportDefault(code);
-  // console.log(code);
 
-  const ast = parser.parse(code, {
-    sourceType: "module",
-    plugins: ["jsx", "typescript"],
-  });
+  try {
+    const ast = parser.parse(code, {
+      sourceType: "module",
+      plugins: ["jsx", "typescript"],
+    });
 
-  const declarations = {};
+    const declarations = {};
 
-  traverse(ast, {
-    ExportNamedDeclaration(path) {
-      if (path.node.declaration) {
-        // Para declarações diretamente na exportação
-        if (path.node.declaration.type === "VariableDeclaration") {
-          path.node.declaration.declarations.forEach((decl) => {
-            if (decl.id.type === "Identifier") {
-              const key = decl.id.name;
-              const value = decl.init
-                ? generate(decl.init, { concise: false }).code
-                : "undefined";
-              declarations[key] = value;
+    traverse(ast, {
+      ExportNamedDeclaration(path) {
+        if (path.node.declaration) {
+          // Para declarações diretamente na exportação
+          if (path.node.declaration.type === "VariableDeclaration") {
+            path.node.declaration.declarations.forEach((decl) => {
+              if (decl.id.type === "Identifier") {
+                const key = decl.id.name;
+                const value = decl.init
+                  ? generate(decl.init, { concise: false }).code
+                  : "undefined";
+                declarations[key] = value;
+              }
+            });
+          } else if (path.node.declaration.type === "FunctionDeclaration") {
+            const { id, params, body, async } = path.node.declaration;
+            if (id) {
+              const arrowFunction = t.arrowFunctionExpression(
+                params,
+                body,
+                async,
+              );
+              path.replaceWith(
+                t.variableDeclaration("const", [
+                  t.variableDeclarator(t.identifier(id.name), arrowFunction),
+                ]),
+              );
+              declarations[id.name] = generate(arrowFunction, {
+                concise: false,
+              }).code;
             }
-          });
-        } else if (path.node.declaration.type === "FunctionDeclaration") {
-          const { id, params, body, async } = path.node.declaration;
-          if (id) {
-            const arrowFunction = t.arrowFunctionExpression(
-              params,
-              body,
-              async,
-            );
-            path.replaceWith(
-              t.variableDeclaration("const", [
-                t.variableDeclarator(t.identifier(id.name), arrowFunction),
-              ]),
-            );
-            declarations[id.name] = generate(arrowFunction, {
-              concise: false,
-            }).code;
           }
+        } else if (path.node.specifiers) {
+          // Para exportações de variáveis ou funções já declaradas
+          path.node.specifiers.forEach((specifier) => {
+            declarations[specifier.exported.name] = specifier.local.name;
+          });
         }
-      } else if (path.node.specifiers) {
-        // Para exportações de variáveis ou funções já declaradas
-        path.node.specifiers.forEach((specifier) => {
-          declarations[specifier.exported.name] = specifier.local.name;
-        });
-      }
-    },
-  });
+      },
+    });
 
-  return declarations;
+    return declarations;
+  } catch (error) {
+    console.log(code);
+    throw new Error(`File: ${filePath}: ${error}`);
+  }
 }
 
 module.exports = transformAndGetFunctionExportDeclarations;
