+++ lib/parsers/wrapCodeInGetFunction.js
@@ -1,48 +1,11 @@
 const parser = require("@babel/parser");
 const traverse = require("@babel/traverse").default;
 const t = require("@babel/types");
-const generate = require("@babel/generator").default;
 const babel = require("@babel/core");
 // const babelPreset = require("@babel/preset-env");
-const reactPreset = require("@babel/preset-react");
+// const reactPreset = require("@babel/preset-react");
 const typescriptPreset = require("@babel/preset-typescript");
-
-/**
- * Converte todas as "function" em arrow functions
- * @param {*} code
- * @returns
- */
-function convertFunctionsToArrow(code) {
-  // Parsear o código para criar a AST
-  const ast = parser.parse(code, {
-    sourceType: "module",
-    plugins: ["jsx", "typescript"], // Adiciona suporte para JSX e TypeScript, se necessário
-  });
-
-  // Transformar todas as funções declaradas e funções anônimas em expressões de arrow
-  traverse(ast, {
-    FunctionDeclaration(path) {
-      const { id, params, body, async } = path.node;
-      const arrowFunction = t.arrowFunctionExpression(params, body, async);
-      arrowFunction.returnType = path.node.returnType; // Preserva o tipo de retorno, se for TypeScript
-      path.replaceWith(
-        t.variableDeclaration("const", [
-          t.variableDeclarator(id, arrowFunction),
-        ]),
-      );
-    },
-    FunctionExpression(path) {
-      const { params, body, async } = path.node;
-      const arrowFunction = t.arrowFunctionExpression(params, body, async);
-      path.replaceWith(arrowFunction);
-    },
-  });
-
-  // Gerar o código JavaScript modificado a partir da AST atualizada
-  return generate(ast, {
-    /* Opções aqui, se necessário */
-  }).code;
-}
+const convertFunctionsToArrow = require("./convertFunctionsToArrow");
 
 /**
  * Esta função vai cobrir o arquivo/code com uma função Get e todas as declarações sendo exportadas
@@ -133,7 +96,8 @@ function wrapCodeInGetFunction(code, filePath) {
     presets: [
       // INFO: Isso estava gerando os helpers no topo do arquivo
       // babelPreset,
-      reactPreset,
+      // INFO: Esse preset estava transformando os arquivos
+      // reactPreset,
       typescriptPreset,
     ],
     code: true,
